Refactor DXC libraries loading logic in D3D12 generator
